restructure our api so we can split it all out into files to make it easier to find things
