+++ src/app/britecharts/settings.ts
@@ -1,0 +1,4 @@
+
+
+export const minimumChartHeight = 50;
+export const minimumChartWidth = 50;
